Update MediaDatabase.java
Previously we exclude audio data from the documents, but we do not include it in "gallery data". This means that audio data is essentially _not included_ under "All Media". A severe implication of this is that audio data cannot be backed up or saved in the same way as documents, images and video.

As of this commit this is completely untested and may not work, depending on if the media viewer supports audio or if that part is separate and only usable in the context of the conversation view. It might work better to put it under the documents view if that is the case, however for this test I would prefer to group it with the saner choice.